Switch to using ->load_classes.
This fixes several problems with classes not being loaded in order, and
no longer requires the tables to be specified in the right order in the
YAML (which is impossible anyway). As a consequence, the table list is
now a hash instead of an array.

In order to use ->load_classes, install a coderef in @INC which does the
work. This coderef actually returns a file with just "1;" in it, but by
that point it's already set up the class properly. This means that to
look for a real .pm we need to use local @INC/do, which is a bit of a
hack but seems to work OK.